refactor: reorganize ui package
