+++ bang_py/ui/__init__.py
@@ -1,0 +1,9 @@
+"""User interface package for the Bang card game.
+
+Provides Qt Quick-based components and utilities for running the game's
+interactive client.
+"""
+
+from .main import BangUI, main
+
+__all__ = ["BangUI", "main"]
