+++ bang_py/ui/qml/Main.qml
